Move shared code to /shared
`/tools` is intended for standalone tools, as stated in the README.
