+++ shared/README.md
@@ -1,0 +1,3 @@
+# Test Infrastructure shared packages
+
+This directory contains shared packages used by other repositories.
